requirements.txt
# AxonVision Camera Calibration - Dependencies
# Python 3.8+ required
# Install with: pip install -r requirements.txt

# Computer vision (includes ArUco/ChArUco detection)
opencv-contrib-python==4.8.1.78

# Numerical computing (pinned for reproducibility)
numpy==1.24.3

# Scientific computing (optimization, spatial transforms)
scipy==1.11.4

# HTTP client for network camera API
requests==2.31.0

# PDF report generation (optional - for --generate-pdf)
reportlab==4.0.9

# Serial communication for INS hardware (VectorNav)
pyserial==3.5

# Qt GUI framework (official Qt for Python)
PySide6==6.6.1

# YAML configuration support
PyYAML==6.0.1
